chore: move breadcrumbs component
this should be in the components directory
